Adding back the compliance and client runs apis
