+++ e2e/cypress/integration/api/iam/12_projects_api_node_ingestion.spec.ts
@@ -1,8 +1,9 @@
 import { describeIfIAMV2p1 } from '../../../support/constants';
 import { uuidv4 } from '../../../support/helpers';
 
-describeIfIAMV2p1('Client Runs Ingestion project tagging', () => {
-  const cypressPrefix = 'test-projects-api-client-runs';
+describeIfIAMV2p1('Ingestion project tagging', () => {
+  const cypressPrefix = 'test-node-ingestion';
+
 
   const projectsWithRule = [
     {
@@ -140,6 +141,12 @@ describeIfIAMV2p1('Client Runs Ingestion project tagging', () => {
       }
     }
   ];
+  const complianceNodeId = uuidv4();
+  const clientRunsNodeId = uuidv4();
+  const reportId = uuidv4();
+  const nodeName = `${cypressPrefix}-${Cypress.moment().format('MMDDYYhhmm')}`;
+  const start = Cypress.moment().utc().subtract(3, 'day').startOf('day').format();
+  const end = Cypress.moment().utc().endOf('day').format();
 
   before(() => {
     cy.cleanupV2IAMObjectsByIDPrefixes(cypressPrefix, ['projects', 'policies']);
@@ -169,14 +176,29 @@ describeIfIAMV2p1('Client Runs Ingestion project tagging', () => {
     });
 
     cy.waitUntilApplyRulesNotRunning(100);
-  });
 
-  after(() => cy.cleanupV2IAMObjectsByIDPrefixes(cypressPrefix, ['projects', 'policies']));
+    // Ingest a InSpec report with attributes that match all the projects
+    cy.fixture('compliance/inspec-report.json').then((report) => {
+      report.organization_name = '75th Rangers';
+      report.source_fqdn = 'example.org';
+      report.environment = 'arctic';
+      report.policy_group = 'red_ring';
+      report.policy_name = 'fire';
+      report.roles = ['backend'];
+      report.chef_tags = ['v3'];
+      report.node_uuid = complianceNodeId;
+      report.node_name = nodeName;
+      report.end_time = Cypress.moment().utc().format();
+      report.report_uuid = reportId;
+      cy.request({
+        headers: { 'api-token': Cypress.env('ADMIN_TOKEN') },
+        method: 'POST',
+        url: '/data-collector/v0',
+        body: report
+      });
+    });
 
-  it('tagging 7 projects with all the attributes on a client runs node', () => {
-    const nodeId = uuidv4();
-
-    // Ingest a node with attribues that match all the projects
+    // Ingest a node with attributes that match all the projects
     cy.fixture('converge/avengers1.json').then((node) => {
       node.organization_name = '75th Rangers';
       node.chef_server_fqdn = 'example.org';
@@ -185,7 +207,7 @@ describeIfIAMV2p1('Client Runs Ingestion project tagging', () => {
       node.policy_name = 'fire';
       node.node.automatic.roles = ['backend'];
       node.node.normal.tags = ['v3'];
-      node.entity_uuid = nodeId;
+      node.entity_uuid = clientRunsNodeId;
       cy.request({
         headers: { 'api-token': Cypress.env('ADMIN_TOKEN') },
         method: 'POST',
@@ -194,33 +216,93 @@ describeIfIAMV2p1('Client Runs Ingestion project tagging', () => {
       });
     });
 
-    // wait for the node to be ingested
-    cy.waitForNodemanagerNode(nodeId, 30);
+    // wait for the report to be ingested
+    cy.waitForNodemanagerNode(complianceNodeId, 30);
+    cy.waitForComplianceNode(complianceNodeId, start, end, 30);
+    // wait for the client run report to be ingested
+    cy.waitForNodemanagerNode(clientRunsNodeId, 30);
+    cy.waitForClientRunsNode(clientRunsNodeId, 30);
+  });
 
-    // Ensure the node is tagged with the correct project
-    projectsWithRule.forEach(projectWithRule => {
-      cy.request({
-        headers: {
-          'api-token': Cypress.env('ADMIN_TOKEN'),
-          projects: projectWithRule.project.id
-        },
-        method: 'GET',
-        url: `/api/v0/cfgmgmt/nodes?pagination.size=10&filter=node_id:${nodeId}`
-      }).then((response) => {
-        expect(response.body).to.have.length(1);
-        expect(response.body[0].id).to.equal(nodeId);
+  after(() => cy.cleanupV2IAMObjectsByIDPrefixes(cypressPrefix, ['projects', 'policies']));
+
+  describe('Node Manager', () => {
+    // Ensure that both the compliance and client run nodes' tags are updated
+    for ( const projectWithRule of projectsWithRule ) {
+      it(`when a project has a rule that matches a node's ${projectWithRule.rule.name},
+      successfully associates that node with the project`, () => {
+        cy.request({
+          headers: {
+            projects: [projectWithRule.project.id],
+            'api-token': Cypress.env('ADMIN_TOKEN')
+          },
+          method: 'POST',
+          url: '/api/v0/nodes/search',
+          body: {
+            order: 'DESC',
+            page: 1,
+            per_page: 100
+          }
+        }).then((response) => {
+          expect(response.body.nodes.length).to.be.greaterThan(0);
+          expect(response.body.nodes.some((node: any) =>
+            node.id === complianceNodeId)).to.equal(true);
+          expect(response.body.nodes.some((node: any) =>
+            node.id === clientRunsNodeId)).to.equal(true);
+        });
       });
-    });
+    }
+  });
 
-    // Delete node
-    cy.request({
-      headers: { 'api-token': Cypress.env('ADMIN_TOKEN') },
-      method: 'POST',
-      url: 'api/v0/ingest/events/chef/node-multiple-deletes',
-      body: {
-        node_ids: [nodeId]
-      },
-      failOnStatusCode: false
-    });
+  describe('Client Runs', () => {
+    for ( const projectWithRule of projectsWithRule ) {
+      it(`when a project has a rule that matches a node's ${projectWithRule.rule.name},
+      successfully associates that node with the project`, () => {
+
+        // Ensure that both the compliance and client run nodes' tags are updated
+        cy.request({
+          headers: {
+            'api-token': Cypress.env('ADMIN_TOKEN'),
+            projects: projectWithRule.project.id
+          },
+          method: 'GET',
+          url: `/api/v0/cfgmgmt/nodes?pagination.size=10&filter=node_id:${clientRunsNodeId}`
+        }).then((response) => {
+          expect(response.body).to.have.length(1);
+          expect(response.body[0].id).to.equal(clientRunsNodeId);
+        });
+      });
+    }
+  });
+
+  describe('Compliance Runs', () => {
+    for ( const projectWithRule of projectsWithRule ) {
+      it(`when a project has a rule that matches a node's ${projectWithRule.rule.name},
+      successfully associates that node with the project`, () => {
+
+        // Ensure the node is tagged with the correct project
+        cy.request({
+          headers: {
+            'api-token': Cypress.env('ADMIN_TOKEN'),
+            projects: [projectWithRule.project.id]
+          },
+          method: 'POST',
+          url: '/api/v0/compliance/reporting/nodes/search',
+          body: {
+            filters: [
+              { type: 'start_time', values: [start]},
+              { type: 'end_time', values: [end]},
+              { type: 'node_id', values: [complianceNodeId]}
+            ],
+            order: 'DESC',
+            page: 1,
+            per_page: 100,
+            sort: 'latest_report.end_time'
+          }
+        }).then((response) => {
+          expect(response.body.nodes).to.have.length(1);
+        });
+      });
+    }
   });
 });
