+++ e2e/cypress/integration/api/iam/13_projects_api_node_update.spec.ts
@@ -1,30 +1,29 @@
 import { describeIfIAMV2p1 } from '../../../support/constants';
 import { uuidv4 } from '../../../support/helpers';
 
-describeIfIAMV2p1('Ingestion project tagging', () => {
-  const cypressPrefix = 'test-node-ingestion';
-
+describeIfIAMV2p1('project update re-tagging', () => {
+  const cypressPrefix = 'test-node-update';
 
   const projectsWithRule = [
-    // {
-    //   project: {
-    //     id: `${cypressPrefix}-project-org-${Cypress.moment().format('MMDDYYhhmm')}`,
-    //     name: 'project org'
-    //   },
-    //   rule: {
-    //     id: 'rule-org',
-    //     name: 'rule CHEF_ORGANIZATION',
-    //     type: 'NODE',
-    //     project_id: `${cypressPrefix}-project-org-${Cypress.moment().format('MMDDYYhhmm')}`,
-    //     conditions: [
-    //       {
-    //         attribute: 'CHEF_ORGANIZATION',
-    //         operator: 'EQUALS',
-    //         values: ['75th Rangers']
-    //       }
-    //     ]
-    //   }
-    // },
+    {
+      project: {
+        id: `${cypressPrefix}-project-org-${Cypress.moment().format('MMDDYYhhmm')}`,
+        name: 'project org'
+      },
+      rule: {
+        id: 'rule-org',
+        name: 'rule CHEF_ORGANIZATION',
+        type: 'NODE',
+        project_id: `${cypressPrefix}-project-org-${Cypress.moment().format('MMDDYYhhmm')}`,
+        conditions: [
+          {
+            attribute: 'CHEF_ORGANIZATION',
+            operator: 'EQUALS',
+            values: ['75th Rangers']
+          }
+        ]
+      }
+    },
     {
       project: {
         id: `${cypressPrefix}-project-chef-server-${Cypress.moment().format('MMDDYYhhmm')}`,
@@ -141,6 +140,7 @@ describeIfIAMV2p1('Ingestion project tagging', () => {
       }
     }
   ];
+
   const complianceNodeId = uuidv4();
   const clientRunsNodeId = uuidv4();
   const reportId = uuidv4();
@@ -150,6 +150,52 @@ describeIfIAMV2p1('Ingestion project tagging', () => {
 
   before(() => {
     cy.cleanupV2IAMObjectsByIDPrefixes(cypressPrefix, ['projects', 'policies']);
+
+    // Ingest a InSpec report with attribues that match all the projects
+    cy.fixture('compliance/inspec-report.json').then((report) => {
+      report.organization_name = '75th Rangers';
+      report.source_fqdn = 'example.org';
+      report.environment = 'arctic';
+      report.policy_group = 'red_ring';
+      report.policy_name = 'fire';
+      report.roles = ['backend'];
+      report.chef_tags = ['v3'];
+      report.node_uuid = complianceNodeId;
+      report.node_name = nodeName;
+      report.end_time = Cypress.moment().utc().format();
+      report.report_uuid = reportId;
+      cy.request({
+        headers: { 'api-token': Cypress.env('ADMIN_TOKEN') },
+        method: 'POST',
+        url: '/data-collector/v0',
+        body: report
+      });
+    });
+
+    // Ingest a node with attribues that match all the projects
+    cy.fixture('converge/avengers1.json').then((node) => {
+      node.organization_name = '75th Rangers';
+      node.chef_server_fqdn = 'example.org';
+      node.node.chef_environment = 'arctic';
+      node.policy_group = 'red_ring';
+      node.policy_name = 'fire';
+      node.node.automatic.roles = ['backend'];
+      node.node.normal.tags = ['v3'];
+      node.entity_uuid = clientRunsNodeId;
+      cy.request({
+        headers: { 'api-token': Cypress.env('ADMIN_TOKEN') },
+        method: 'POST',
+        url: '/data-collector/v0',
+        body: node
+      });
+    });
+
+    // wait for the report to be ingested
+    cy.waitForNodemanagerNode(complianceNodeId, 30);
+    cy.waitForComplianceNode(complianceNodeId, start, end, 30);
+    // wait for the client run report to be ingested
+    cy.waitForNodemanagerNode(clientRunsNodeId, 30);
+    cy.waitForClientRunsNode(clientRunsNodeId, 30);
 
     // create the projects with one rule
     projectsWithRule.forEach(projectWithRule => {
@@ -176,59 +222,16 @@ describeIfIAMV2p1('Ingestion project tagging', () => {
     });
 
     cy.waitUntilApplyRulesNotRunning(100);
-
-    // Ingest a InSpec report with attributes that match all the projects
-    cy.fixture('compliance/inspec-report.json').then((report) => {
-      report.organization_name = '75th Rangers';
-      report.source_fqdn = 'example.org';
-      report.environment = 'arctic';
-      report.policy_group = 'red_ring';
-      report.policy_name = 'fire';
-      report.roles = ['backend'];
-      report.chef_tags = ['v3'];
-      report.node_uuid = complianceNodeId;
-      report.node_name = nodeName;
-      report.end_time = Cypress.moment().utc().format();
-      report.report_uuid = reportId;
-      cy.request({
-        headers: { 'api-token': Cypress.env('ADMIN_TOKEN') },
-        method: 'POST',
-        url: '/data-collector/v0',
-        body: report
-      });
-    });
-
-    // Ingest a node with attributes that match all the projects
-    cy.fixture('converge/avengers1.json').then((node) => {
-      node.organization_name = '75th Rangers';
-      node.chef_server_fqdn = 'example.org';
-      node.node.chef_environment = 'arctic';
-      node.policy_group = 'red_ring';
-      node.policy_name = 'fire';
-      node.node.automatic.roles = ['backend'];
-      node.node.normal.tags = ['v3'];
-      node.entity_uuid = clientRunsNodeId;
-      cy.request({
-        headers: { 'api-token': Cypress.env('ADMIN_TOKEN') },
-        method: 'POST',
-        url: '/data-collector/v0',
-        body: node
-      });
-    });
-
-    // wait for the report to be ingested
-    cy.waitForNodemanagerNode(complianceNodeId, 30);
-    // wait for the client run report to be ingested
-    cy.waitForNodemanagerNode(clientRunsNodeId, 30);
   });
 
   after(() => cy.cleanupV2IAMObjectsByIDPrefixes(cypressPrefix, ['projects', 'policies']));
 
   describe('Node Manager', () => {
-    // Ensure that both the compliance and client run nodes' tags are updated
     for ( const projectWithRule of projectsWithRule ) {
       it(`when a project has a rule that matches a node's ${projectWithRule.rule.name},
       successfully associates that node with the project`, () => {
+
+        // Ensure that both the compliance and client run nodes' tags are updated
         cy.request({
           headers: {
             projects: [projectWithRule.project.id],
